+++ tests/rest/client/sliding_sync/test_sliding_sync.py
@@ -932,2446 +932,6 @@ class SlidingSyncTestCase(SlidingSyncBase):
             response_body["lists"]["foo-list"],
         )
 
-    def test_rooms_meta_when_joined(self) -> None:
-        """
-        Test that the `rooms` `name` and `avatar` are included in the response and
-        reflect the current state of the room when the user is joined to the room.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "name": "my super room",
-            },
-        )
-        # Set the room avatar URL
-        self.helper.send_state(
-            room_id1,
-            EventTypes.RoomAvatar,
-            {"url": "mxc://DUMMY_MEDIA_ID"},
-            tok=user2_tok,
-        )
-
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # Reflect the current state of the room
-        self.assertEqual(
-            response_body["rooms"][room_id1]["name"],
-            "my super room",
-            response_body["rooms"][room_id1],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["avatar"],
-            "mxc://DUMMY_MEDIA_ID",
-            response_body["rooms"][room_id1],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["joined_count"],
-            2,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["invited_count"],
-            0,
-        )
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("is_dm"),
-        )
-
-    def test_rooms_meta_when_invited(self) -> None:
-        """
-        Test that the `rooms` `name` and `avatar` are included in the response and
-        reflect the current state of the room when the user is invited to the room.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "name": "my super room",
-            },
-        )
-        # Set the room avatar URL
-        self.helper.send_state(
-            room_id1,
-            EventTypes.RoomAvatar,
-            {"url": "mxc://DUMMY_MEDIA_ID"},
-            tok=user2_tok,
-        )
-
-        # User1 is invited to the room
-        self.helper.invite(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-
-        # Update the room name after user1 has left
-        self.helper.send_state(
-            room_id1,
-            EventTypes.Name,
-            {"name": "my super duper room"},
-            tok=user2_tok,
-        )
-        # Update the room avatar URL after user1 has left
-        self.helper.send_state(
-            room_id1,
-            EventTypes.RoomAvatar,
-            {"url": "mxc://UPDATED_DUMMY_MEDIA_ID"},
-            tok=user2_tok,
-        )
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # This should still reflect the current state of the room even when the user is
-        # invited.
-        self.assertEqual(
-            response_body["rooms"][room_id1]["name"],
-            "my super duper room",
-            response_body["rooms"][room_id1],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["avatar"],
-            "mxc://UPDATED_DUMMY_MEDIA_ID",
-            response_body["rooms"][room_id1],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["joined_count"],
-            1,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["invited_count"],
-            1,
-        )
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("is_dm"),
-        )
-
-    def test_rooms_meta_when_banned(self) -> None:
-        """
-        Test that the `rooms` `name` and `avatar` reflect the state of the room when the
-        user was banned (do not leak current state).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "name": "my super room",
-            },
-        )
-        # Set the room avatar URL
-        self.helper.send_state(
-            room_id1,
-            EventTypes.RoomAvatar,
-            {"url": "mxc://DUMMY_MEDIA_ID"},
-            tok=user2_tok,
-        )
-
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-        self.helper.ban(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-
-        # Update the room name after user1 has left
-        self.helper.send_state(
-            room_id1,
-            EventTypes.Name,
-            {"name": "my super duper room"},
-            tok=user2_tok,
-        )
-        # Update the room avatar URL after user1 has left
-        self.helper.send_state(
-            room_id1,
-            EventTypes.RoomAvatar,
-            {"url": "mxc://UPDATED_DUMMY_MEDIA_ID"},
-            tok=user2_tok,
-        )
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # Reflect the state of the room at the time of leaving
-        self.assertEqual(
-            response_body["rooms"][room_id1]["name"],
-            "my super room",
-            response_body["rooms"][room_id1],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["avatar"],
-            "mxc://DUMMY_MEDIA_ID",
-            response_body["rooms"][room_id1],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["joined_count"],
-            # FIXME: The actual number should be "1" (user2) but we currently don't
-            # support this for rooms where the user has left/been banned.
-            0,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["invited_count"],
-            0,
-        )
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("is_dm"),
-        )
-
-    def test_rooms_meta_heroes(self) -> None:
-        """
-        Test that the `rooms` `heroes` are included in the response when the room
-        doesn't have a room name set.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user3_id = self.register_user("user3", "pass")
-        _user3_tok = self.login(user3_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "name": "my super room",
-            },
-        )
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-        # User3 is invited
-        self.helper.invite(room_id1, src=user2_id, targ=user3_id, tok=user2_tok)
-
-        room_id2 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                # No room name set so that `heroes` is populated
-                #
-                # "name": "my super room2",
-            },
-        )
-        self.helper.join(room_id2, user1_id, tok=user1_tok)
-        # User3 is invited
-        self.helper.invite(room_id2, src=user2_id, targ=user3_id, tok=user2_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # Room1 has a name so we shouldn't see any `heroes` which the client would use
-        # the calculate the room name themselves.
-        self.assertEqual(
-            response_body["rooms"][room_id1]["name"],
-            "my super room",
-            response_body["rooms"][room_id1],
-        )
-        self.assertIsNone(response_body["rooms"][room_id1].get("heroes"))
-        self.assertEqual(
-            response_body["rooms"][room_id1]["joined_count"],
-            2,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["invited_count"],
-            1,
-        )
-
-        # Room2 doesn't have a name so we should see `heroes` populated
-        self.assertIsNone(response_body["rooms"][room_id2].get("name"))
-        self.assertCountEqual(
-            [
-                hero["user_id"]
-                for hero in response_body["rooms"][room_id2].get("heroes", [])
-            ],
-            # Heroes shouldn't include the user themselves (we shouldn't see user1)
-            [user2_id, user3_id],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id2]["joined_count"],
-            2,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id2]["invited_count"],
-            1,
-        )
-
-        # We didn't request any state so we shouldn't see any `required_state`
-        self.assertIsNone(response_body["rooms"][room_id1].get("required_state"))
-        self.assertIsNone(response_body["rooms"][room_id2].get("required_state"))
-
-    def test_rooms_meta_heroes_max(self) -> None:
-        """
-        Test that the `rooms` `heroes` only includes the first 5 users (not including
-        yourself).
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user3_id = self.register_user("user3", "pass")
-        user3_tok = self.login(user3_id, "pass")
-        user4_id = self.register_user("user4", "pass")
-        user4_tok = self.login(user4_id, "pass")
-        user5_id = self.register_user("user5", "pass")
-        user5_tok = self.login(user5_id, "pass")
-        user6_id = self.register_user("user6", "pass")
-        user6_tok = self.login(user6_id, "pass")
-        user7_id = self.register_user("user7", "pass")
-        user7_tok = self.login(user7_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                # No room name set so that `heroes` is populated
-                #
-                # "name": "my super room",
-            },
-        )
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-        self.helper.join(room_id1, user3_id, tok=user3_tok)
-        self.helper.join(room_id1, user4_id, tok=user4_tok)
-        self.helper.join(room_id1, user5_id, tok=user5_tok)
-        self.helper.join(room_id1, user6_id, tok=user6_tok)
-        self.helper.join(room_id1, user7_id, tok=user7_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # Room2 doesn't have a name so we should see `heroes` populated
-        self.assertIsNone(response_body["rooms"][room_id1].get("name"))
-        self.assertCountEqual(
-            [
-                hero["user_id"]
-                for hero in response_body["rooms"][room_id1].get("heroes", [])
-            ],
-            # Heroes should be the first 5 users in the room (excluding the user
-            # themselves, we shouldn't see `user1`)
-            [user2_id, user3_id, user4_id, user5_id, user6_id],
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["joined_count"],
-            7,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["invited_count"],
-            0,
-        )
-
-        # We didn't request any state so we shouldn't see any `required_state`
-        self.assertIsNone(response_body["rooms"][room_id1].get("required_state"))
-
-    def test_rooms_meta_heroes_when_banned(self) -> None:
-        """
-        Test that the `rooms` `heroes` are included in the response when the room
-        doesn't have a room name set but doesn't leak information past their ban.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user3_id = self.register_user("user3", "pass")
-        _user3_tok = self.login(user3_id, "pass")
-        user4_id = self.register_user("user4", "pass")
-        user4_tok = self.login(user4_id, "pass")
-        user5_id = self.register_user("user5", "pass")
-        _user5_tok = self.login(user5_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                # No room name set so that `heroes` is populated
-                #
-                # "name": "my super room",
-            },
-        )
-        # User1 joins the room
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-        # User3 is invited
-        self.helper.invite(room_id1, src=user2_id, targ=user3_id, tok=user2_tok)
-
-        # User1 is banned from the room
-        self.helper.ban(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-
-        # User4 joins the room after user1 is banned
-        self.helper.join(room_id1, user4_id, tok=user4_tok)
-        # User5 is invited after user1 is banned
-        self.helper.invite(room_id1, src=user2_id, targ=user5_id, tok=user2_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # Room2 doesn't have a name so we should see `heroes` populated
-        self.assertIsNone(response_body["rooms"][room_id1].get("name"))
-        self.assertCountEqual(
-            [
-                hero["user_id"]
-                for hero in response_body["rooms"][room_id1].get("heroes", [])
-            ],
-            # Heroes shouldn't include the user themselves (we shouldn't see user1). We
-            # also shouldn't see user4 since they joined after user1 was banned.
-            #
-            # FIXME: The actual result should be `[user2_id, user3_id]` but we currently
-            # don't support this for rooms where the user has left/been banned.
-            [],
-        )
-
-        self.assertEqual(
-            response_body["rooms"][room_id1]["joined_count"],
-            # FIXME: The actual number should be "1" (user2) but we currently don't
-            # support this for rooms where the user has left/been banned.
-            0,
-        )
-        self.assertEqual(
-            response_body["rooms"][room_id1]["invited_count"],
-            # We shouldn't see user5 since they were invited after user1 was banned.
-            #
-            # FIXME: The actual number should be "1" (user3) but we currently don't
-            # support this for rooms where the user has left/been banned.
-            0,
-        )
-
-    def test_rooms_limited_initial_sync(self) -> None:
-        """
-        Test that we mark `rooms` as `limited=True` when we saturate the `timeline_limit`
-        on initial sync.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity1", tok=user2_tok)
-        self.helper.send(room_id1, "activity2", tok=user2_tok)
-        event_response3 = self.helper.send(room_id1, "activity3", tok=user2_tok)
-        event_pos3 = self.get_success(
-            self.store.get_position_for_event(event_response3["event_id"])
-        )
-        event_response4 = self.helper.send(room_id1, "activity4", tok=user2_tok)
-        event_pos4 = self.get_success(
-            self.store.get_position_for_event(event_response4["event_id"])
-        )
-        event_response5 = self.helper.send(room_id1, "activity5", tok=user2_tok)
-        user1_join_response = self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 3,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # We expect to saturate the `timeline_limit` (there are more than 3 messages in the room)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            True,
-            response_body["rooms"][room_id1],
-        )
-        # Check to make sure the latest events are returned
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                event_response4["event_id"],
-                event_response5["event_id"],
-                user1_join_response["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-
-        # Check to make sure the `prev_batch` points at the right place
-        prev_batch_token = self.get_success(
-            StreamToken.from_string(
-                self.store, response_body["rooms"][room_id1]["prev_batch"]
-            )
-        )
-        prev_batch_room_stream_token_serialized = self.get_success(
-            prev_batch_token.room_key.to_string(self.store)
-        )
-        # If we use the `prev_batch` token to look backwards, we should see `event3`
-        # next so make sure the token encompasses it
-        self.assertEqual(
-            event_pos3.persisted_after(prev_batch_token.room_key),
-            False,
-            f"`prev_batch` token {prev_batch_room_stream_token_serialized} should be >= event_pos3={self.get_success(event_pos3.to_room_stream_token().to_string(self.store))}",
-        )
-        # If we use the `prev_batch` token to look backwards, we shouldn't see `event4`
-        # anymore since it was just returned in this response.
-        self.assertEqual(
-            event_pos4.persisted_after(prev_batch_token.room_key),
-            True,
-            f"`prev_batch` token {prev_batch_room_stream_token_serialized} should be < event_pos4={self.get_success(event_pos4.to_room_stream_token().to_string(self.store))}",
-        )
-
-        # With no `from_token` (initial sync), it's all historical since there is no
-        # "live" range
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            0,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_rooms_not_limited_initial_sync(self) -> None:
-        """
-        Test that we mark `rooms` as `limited=False` when there are no more events to
-        paginate to.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity1", tok=user2_tok)
-        self.helper.send(room_id1, "activity2", tok=user2_tok)
-        self.helper.send(room_id1, "activity3", tok=user2_tok)
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        # Make the Sliding Sync request
-        timeline_limit = 100
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": timeline_limit,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # The timeline should be `limited=False` because we have all of the events (no
-        # more to paginate to)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            False,
-            response_body["rooms"][room_id1],
-        )
-        expected_number_of_events = 9
-        # We're just looking to make sure we got all of the events before hitting the `timeline_limit`
-        self.assertEqual(
-            len(response_body["rooms"][room_id1]["timeline"]),
-            expected_number_of_events,
-            response_body["rooms"][room_id1]["timeline"],
-        )
-        self.assertLessEqual(expected_number_of_events, timeline_limit)
-
-        # With no `from_token` (initial sync), it's all historical since there is no
-        # "live" token range.
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            0,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_rooms_incremental_sync(self) -> None:
-        """
-        Test `rooms` data during an incremental sync after an initial sync.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-        self.helper.send(room_id1, "activity before initial sync1", tok=user2_tok)
-
-        # Make an initial Sliding Sync request to grab a token. This is also a sanity
-        # check that we can go from initial to incremental sync.
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 3,
-                }
-            }
-        }
-        _, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        # Send some events but don't send enough to saturate the `timeline_limit`.
-        # We want to later test that we only get the new events since the `next_pos`
-        event_response2 = self.helper.send(room_id1, "activity after2", tok=user2_tok)
-        event_response3 = self.helper.send(room_id1, "activity after3", tok=user2_tok)
-
-        # Make an incremental Sliding Sync request (what we're trying to test)
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        # We only expect to see the new events since the last sync which isn't enough to
-        # fill up the `timeline_limit`.
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            False,
-            f'Our `timeline_limit` was {sync_body["lists"]["foo-list"]["timeline_limit"]} '
-            + f'and {len(response_body["rooms"][room_id1]["timeline"])} events were returned in the timeline. '
-            + str(response_body["rooms"][room_id1]),
-        )
-        # Check to make sure the latest events are returned
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                event_response2["event_id"],
-                event_response3["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-
-        # All events are "live"
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            2,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_rooms_bump_stamp(self) -> None:
-        """
-        Test that `bump_stamp` is present and pointing to relevant events.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        room_id1 = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-        )
-        event_response1 = message_response = self.helper.send(
-            room_id1, "message in room1", tok=user1_tok
-        )
-        event_pos1 = self.get_success(
-            self.store.get_position_for_event(event_response1["event_id"])
-        )
-        room_id2 = self.helper.create_room_as(
-            user1_id,
-            tok=user1_tok,
-        )
-        send_response2 = self.helper.send(room_id2, "message in room2", tok=user1_tok)
-        event_pos2 = self.get_success(
-            self.store.get_position_for_event(send_response2["event_id"])
-        )
-
-        # Send a reaction in room1 but it shouldn't affect the `bump_stamp`
-        # because reactions are not part of the `DEFAULT_BUMP_EVENT_TYPES`
-        self.helper.send_event(
-            room_id1,
-            type=EventTypes.Reaction,
-            content={
-                "m.relates_to": {
-                    "event_id": message_response["event_id"],
-                    "key": "👍",
-                    "rel_type": "m.annotation",
-                }
-            },
-            tok=user1_tok,
-        )
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 100,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # Make sure it has the foo-list we requested
-        self.assertListEqual(
-            list(response_body["lists"].keys()),
-            ["foo-list"],
-            response_body["lists"].keys(),
-        )
-
-        # Make sure the list includes the rooms in the right order
-        self.assertListEqual(
-            list(response_body["lists"]["foo-list"]["ops"]),
-            [
-                {
-                    "op": "SYNC",
-                    "range": [0, 1],
-                    # room1 sorts before room2 because it has the latest event (the
-                    # reaction)
-                    "room_ids": [room_id1, room_id2],
-                }
-            ],
-            response_body["lists"]["foo-list"],
-        )
-
-        # The `bump_stamp` for room1 should point at the latest message (not the
-        # reaction since it's not one of the `DEFAULT_BUMP_EVENT_TYPES`)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["bump_stamp"],
-            event_pos1.stream,
-            response_body["rooms"][room_id1],
-        )
-
-        # The `bump_stamp` for room2 should point at the latest message
-        self.assertEqual(
-            response_body["rooms"][room_id2]["bump_stamp"],
-            event_pos2.stream,
-            response_body["rooms"][room_id2],
-        )
-
-    def test_rooms_bump_stamp_backfill(self) -> None:
-        """
-        Test that `bump_stamp` ignores backfilled events, i.e. events with a
-        negative stream ordering.
-        """
-
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create a remote room
-        creator = "@user:other"
-        room_id = "!foo:other"
-        shared_kwargs = {
-            "room_id": room_id,
-            "room_version": "10",
-        }
-
-        create_tuple = self.get_success(
-            create_event(
-                self.hs,
-                prev_event_ids=[],
-                type=EventTypes.Create,
-                state_key="",
-                sender=creator,
-                **shared_kwargs,
-            )
-        )
-        creator_tuple = self.get_success(
-            create_event(
-                self.hs,
-                prev_event_ids=[create_tuple[0].event_id],
-                auth_event_ids=[create_tuple[0].event_id],
-                type=EventTypes.Member,
-                state_key=creator,
-                content={"membership": Membership.JOIN},
-                sender=creator,
-                **shared_kwargs,
-            )
-        )
-        # We add a message event as a valid "bump type"
-        msg_tuple = self.get_success(
-            create_event(
-                self.hs,
-                prev_event_ids=[creator_tuple[0].event_id],
-                auth_event_ids=[create_tuple[0].event_id],
-                type=EventTypes.Message,
-                content={"body": "foo", "msgtype": "m.text"},
-                sender=creator,
-                **shared_kwargs,
-            )
-        )
-        invite_tuple = self.get_success(
-            create_event(
-                self.hs,
-                prev_event_ids=[msg_tuple[0].event_id],
-                auth_event_ids=[create_tuple[0].event_id, creator_tuple[0].event_id],
-                type=EventTypes.Member,
-                state_key=user1_id,
-                content={"membership": Membership.INVITE},
-                sender=creator,
-                **shared_kwargs,
-            )
-        )
-
-        remote_events_and_contexts = [
-            create_tuple,
-            creator_tuple,
-            msg_tuple,
-            invite_tuple,
-        ]
-
-        # Ensure the local HS knows the room version
-        self.get_success(
-            self.store.store_room(room_id, creator, False, RoomVersions.V10)
-        )
-
-        # Persist these events as backfilled events.
-        persistence = self.hs.get_storage_controllers().persistence
-        assert persistence is not None
-
-        for event, context in remote_events_and_contexts:
-            self.get_success(persistence.persist_event(event, context, backfilled=True))
-
-        # Now we join the local user to the room
-        join_tuple = self.get_success(
-            create_event(
-                self.hs,
-                prev_event_ids=[invite_tuple[0].event_id],
-                auth_event_ids=[create_tuple[0].event_id, invite_tuple[0].event_id],
-                type=EventTypes.Member,
-                state_key=user1_id,
-                content={"membership": Membership.JOIN},
-                sender=user1_id,
-                **shared_kwargs,
-            )
-        )
-        self.get_success(persistence.persist_event(*join_tuple))
-
-        # Doing an SS request should return a positive `bump_stamp`, even though
-        # the only event that matches the bump types has as negative stream
-        # ordering.
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 5,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        self.assertGreater(response_body["rooms"][room_id]["bump_stamp"], 0)
-
-    def test_rooms_newly_joined_incremental_sync(self) -> None:
-        """
-        Test that when we make an incremental sync with a `newly_joined` `rooms`, we are
-        able to see some historical events before the `from_token`.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity before token1", tok=user2_tok)
-        event_response2 = self.helper.send(
-            room_id1, "activity before token2", tok=user2_tok
-        )
-
-        # The `timeline_limit` is set to 4 so we can at least see one historical event
-        # before the `from_token`. We should see historical events because this is a
-        # `newly_joined` room.
-        timeline_limit = 4
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": timeline_limit,
-                }
-            }
-        }
-        _, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        # Join the room after the `from_token` which will make us consider this room as
-        # `newly_joined`.
-        user1_join_response = self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        # Send some events but don't send enough to saturate the `timeline_limit`.
-        # We want to later test that we only get the new events since the `next_pos`
-        event_response3 = self.helper.send(
-            room_id1, "activity after token3", tok=user2_tok
-        )
-        event_response4 = self.helper.send(
-            room_id1, "activity after token4", tok=user2_tok
-        )
-
-        # Make an incremental Sliding Sync request (what we're trying to test)
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        # We should see the new events and the rest should be filled with historical
-        # events which will make us `limited=True` since there are more to paginate to.
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            True,
-            f"Our `timeline_limit` was {timeline_limit} "
-            + f'and {len(response_body["rooms"][room_id1]["timeline"])} events were returned in the timeline. '
-            + str(response_body["rooms"][room_id1]),
-        )
-        # Check to make sure that the "live" and historical events are returned
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                event_response2["event_id"],
-                user1_join_response["event_id"],
-                event_response3["event_id"],
-                event_response4["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-
-        # Only events after the `from_token` are "live" (join, event3, event4)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            3,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_rooms_invite_shared_history_initial_sync(self) -> None:
-        """
-        Test that `rooms` we are invited to have some stripped `invite_state` during an
-        initial sync.
-
-        This is an `invite` room so we should only have `stripped_state` (no `timeline`)
-        but we also shouldn't see any timeline events because the history visiblity is
-        `shared` and we haven't joined the room yet.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user1 = UserID.from_string(user1_id)
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user2 = UserID.from_string(user2_id)
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        # Ensure we're testing with a room with `shared` history visibility which means
-        # history visible until you actually join the room.
-        history_visibility_response = self.helper.get_state(
-            room_id1, EventTypes.RoomHistoryVisibility, tok=user2_tok
-        )
-        self.assertEqual(
-            history_visibility_response.get("history_visibility"),
-            HistoryVisibility.SHARED,
-        )
-
-        self.helper.send(room_id1, "activity before1", tok=user2_tok)
-        self.helper.send(room_id1, "activity before2", tok=user2_tok)
-        self.helper.invite(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity after3", tok=user2_tok)
-        self.helper.send(room_id1, "activity after4", tok=user2_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 3,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # `timeline` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("timeline"),
-            response_body["rooms"][room_id1],
-        )
-        # `num_live` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("num_live"),
-            response_body["rooms"][room_id1],
-        )
-        # `limited` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("limited"),
-            response_body["rooms"][room_id1],
-        )
-        # `prev_batch` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("prev_batch"),
-            response_body["rooms"][room_id1],
-        )
-        # `required_state` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("required_state"),
-            response_body["rooms"][room_id1],
-        )
-        # We should have some `stripped_state` so the potential joiner can identify the
-        # room (we don't care about the order).
-        self.assertCountEqual(
-            response_body["rooms"][room_id1]["invite_state"],
-            [
-                {
-                    "content": {"creator": user2_id, "room_version": "10"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.create",
-                },
-                {
-                    "content": {"join_rule": "public"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.join_rules",
-                },
-                {
-                    "content": {"displayname": user2.localpart, "membership": "join"},
-                    "sender": user2_id,
-                    "state_key": user2_id,
-                    "type": "m.room.member",
-                },
-                {
-                    "content": {"displayname": user1.localpart, "membership": "invite"},
-                    "sender": user2_id,
-                    "state_key": user1_id,
-                    "type": "m.room.member",
-                },
-            ],
-            response_body["rooms"][room_id1]["invite_state"],
-        )
-
-    def test_rooms_invite_shared_history_incremental_sync(self) -> None:
-        """
-        Test that `rooms` we are invited to have some stripped `invite_state` during an
-        incremental sync.
-
-        This is an `invite` room so we should only have `stripped_state` (no `timeline`)
-        but we also shouldn't see any timeline events because the history visiblity is
-        `shared` and we haven't joined the room yet.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user1 = UserID.from_string(user1_id)
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user2 = UserID.from_string(user2_id)
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        # Ensure we're testing with a room with `shared` history visibility which means
-        # history visible until you actually join the room.
-        history_visibility_response = self.helper.get_state(
-            room_id1, EventTypes.RoomHistoryVisibility, tok=user2_tok
-        )
-        self.assertEqual(
-            history_visibility_response.get("history_visibility"),
-            HistoryVisibility.SHARED,
-        )
-
-        self.helper.send(room_id1, "activity before invite1", tok=user2_tok)
-        self.helper.send(room_id1, "activity before invite2", tok=user2_tok)
-        self.helper.invite(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity after invite3", tok=user2_tok)
-        self.helper.send(room_id1, "activity after invite4", tok=user2_tok)
-
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 3,
-                }
-            }
-        }
-        _, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        self.helper.send(room_id1, "activity after token5", tok=user2_tok)
-        self.helper.send(room_id1, "activity after toekn6", tok=user2_tok)
-
-        # Make the Sliding Sync request
-        response_body, from_token = self.do_sync(
-            sync_body, since=from_token, tok=user1_tok
-        )
-
-        # `timeline` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("timeline"),
-            response_body["rooms"][room_id1],
-        )
-        # `num_live` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("num_live"),
-            response_body["rooms"][room_id1],
-        )
-        # `limited` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("limited"),
-            response_body["rooms"][room_id1],
-        )
-        # `prev_batch` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("prev_batch"),
-            response_body["rooms"][room_id1],
-        )
-        # `required_state` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("required_state"),
-            response_body["rooms"][room_id1],
-        )
-        # We should have some `stripped_state` so the potential joiner can identify the
-        # room (we don't care about the order).
-        self.assertCountEqual(
-            response_body["rooms"][room_id1]["invite_state"],
-            [
-                {
-                    "content": {"creator": user2_id, "room_version": "10"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.create",
-                },
-                {
-                    "content": {"join_rule": "public"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.join_rules",
-                },
-                {
-                    "content": {"displayname": user2.localpart, "membership": "join"},
-                    "sender": user2_id,
-                    "state_key": user2_id,
-                    "type": "m.room.member",
-                },
-                {
-                    "content": {"displayname": user1.localpart, "membership": "invite"},
-                    "sender": user2_id,
-                    "state_key": user1_id,
-                    "type": "m.room.member",
-                },
-            ],
-            response_body["rooms"][room_id1]["invite_state"],
-        )
-
-    def test_rooms_invite_world_readable_history_initial_sync(self) -> None:
-        """
-        Test that `rooms` we are invited to have some stripped `invite_state` during an
-        initial sync.
-
-        This is an `invite` room so we should only have `stripped_state` (no `timeline`)
-        but depending on the semantics we decide, we could potentially see some
-        historical events before/after the `from_token` because the history is
-        `world_readable`. Same situation for events after the `from_token` if the
-        history visibility was set to `invited`.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user1 = UserID.from_string(user1_id)
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user2 = UserID.from_string(user2_id)
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "preset": "public_chat",
-                "initial_state": [
-                    {
-                        "content": {
-                            "history_visibility": HistoryVisibility.WORLD_READABLE
-                        },
-                        "state_key": "",
-                        "type": EventTypes.RoomHistoryVisibility,
-                    }
-                ],
-            },
-        )
-        # Ensure we're testing with a room with `world_readable` history visibility
-        # which means events are visible to anyone even without membership.
-        history_visibility_response = self.helper.get_state(
-            room_id1, EventTypes.RoomHistoryVisibility, tok=user2_tok
-        )
-        self.assertEqual(
-            history_visibility_response.get("history_visibility"),
-            HistoryVisibility.WORLD_READABLE,
-        )
-
-        self.helper.send(room_id1, "activity before1", tok=user2_tok)
-        self.helper.send(room_id1, "activity before2", tok=user2_tok)
-        self.helper.invite(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity after3", tok=user2_tok)
-        self.helper.send(room_id1, "activity after4", tok=user2_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    # Large enough to see the latest events and before the invite
-                    "timeline_limit": 4,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # `timeline` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("timeline"),
-            response_body["rooms"][room_id1],
-        )
-        # `num_live` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("num_live"),
-            response_body["rooms"][room_id1],
-        )
-        # `limited` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("limited"),
-            response_body["rooms"][room_id1],
-        )
-        # `prev_batch` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("prev_batch"),
-            response_body["rooms"][room_id1],
-        )
-        # `required_state` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("required_state"),
-            response_body["rooms"][room_id1],
-        )
-        # We should have some `stripped_state` so the potential joiner can identify the
-        # room (we don't care about the order).
-        self.assertCountEqual(
-            response_body["rooms"][room_id1]["invite_state"],
-            [
-                {
-                    "content": {"creator": user2_id, "room_version": "10"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.create",
-                },
-                {
-                    "content": {"join_rule": "public"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.join_rules",
-                },
-                {
-                    "content": {"displayname": user2.localpart, "membership": "join"},
-                    "sender": user2_id,
-                    "state_key": user2_id,
-                    "type": "m.room.member",
-                },
-                {
-                    "content": {"displayname": user1.localpart, "membership": "invite"},
-                    "sender": user2_id,
-                    "state_key": user1_id,
-                    "type": "m.room.member",
-                },
-            ],
-            response_body["rooms"][room_id1]["invite_state"],
-        )
-
-    def test_rooms_invite_world_readable_history_incremental_sync(self) -> None:
-        """
-        Test that `rooms` we are invited to have some stripped `invite_state` during an
-        incremental sync.
-
-        This is an `invite` room so we should only have `stripped_state` (no `timeline`)
-        but depending on the semantics we decide, we could potentially see some
-        historical events before/after the `from_token` because the history is
-        `world_readable`. Same situation for events after the `from_token` if the
-        history visibility was set to `invited`.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user1 = UserID.from_string(user1_id)
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-        user2 = UserID.from_string(user2_id)
-
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "preset": "public_chat",
-                "initial_state": [
-                    {
-                        "content": {
-                            "history_visibility": HistoryVisibility.WORLD_READABLE
-                        },
-                        "state_key": "",
-                        "type": EventTypes.RoomHistoryVisibility,
-                    }
-                ],
-            },
-        )
-        # Ensure we're testing with a room with `world_readable` history visibility
-        # which means events are visible to anyone even without membership.
-        history_visibility_response = self.helper.get_state(
-            room_id1, EventTypes.RoomHistoryVisibility, tok=user2_tok
-        )
-        self.assertEqual(
-            history_visibility_response.get("history_visibility"),
-            HistoryVisibility.WORLD_READABLE,
-        )
-
-        self.helper.send(room_id1, "activity before invite1", tok=user2_tok)
-        self.helper.send(room_id1, "activity before invite2", tok=user2_tok)
-        self.helper.invite(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity after invite3", tok=user2_tok)
-        self.helper.send(room_id1, "activity after invite4", tok=user2_tok)
-
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    # Large enough to see the latest events and before the invite
-                    "timeline_limit": 4,
-                }
-            }
-        }
-        _, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        self.helper.send(room_id1, "activity after token5", tok=user2_tok)
-        self.helper.send(room_id1, "activity after toekn6", tok=user2_tok)
-
-        # Make the incremental Sliding Sync request
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        # `timeline` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("timeline"),
-            response_body["rooms"][room_id1],
-        )
-        # `num_live` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("num_live"),
-            response_body["rooms"][room_id1],
-        )
-        # `limited` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("limited"),
-            response_body["rooms"][room_id1],
-        )
-        # `prev_batch` is omitted for `invite` rooms with `stripped_state` (no timeline anyway)
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("prev_batch"),
-            response_body["rooms"][room_id1],
-        )
-        # `required_state` is omitted for `invite` rooms with `stripped_state`
-        self.assertIsNone(
-            response_body["rooms"][room_id1].get("required_state"),
-            response_body["rooms"][room_id1],
-        )
-        # We should have some `stripped_state` so the potential joiner can identify the
-        # room (we don't care about the order).
-        self.assertCountEqual(
-            response_body["rooms"][room_id1]["invite_state"],
-            [
-                {
-                    "content": {"creator": user2_id, "room_version": "10"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.create",
-                },
-                {
-                    "content": {"join_rule": "public"},
-                    "sender": user2_id,
-                    "state_key": "",
-                    "type": "m.room.join_rules",
-                },
-                {
-                    "content": {"displayname": user2.localpart, "membership": "join"},
-                    "sender": user2_id,
-                    "state_key": user2_id,
-                    "type": "m.room.member",
-                },
-                {
-                    "content": {"displayname": user1.localpart, "membership": "invite"},
-                    "sender": user2_id,
-                    "state_key": user1_id,
-                    "type": "m.room.member",
-                },
-            ],
-            response_body["rooms"][room_id1]["invite_state"],
-        )
-
-    def test_rooms_ban_initial_sync(self) -> None:
-        """
-        Test that `rooms` we are banned from in an intial sync only allows us to see
-        timeline events up to the ban event.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity before1", tok=user2_tok)
-        self.helper.send(room_id1, "activity before2", tok=user2_tok)
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        event_response3 = self.helper.send(room_id1, "activity after3", tok=user2_tok)
-        event_response4 = self.helper.send(room_id1, "activity after4", tok=user2_tok)
-        user1_ban_response = self.helper.ban(
-            room_id1, src=user2_id, targ=user1_id, tok=user2_tok
-        )
-
-        self.helper.send(room_id1, "activity after5", tok=user2_tok)
-        self.helper.send(room_id1, "activity after6", tok=user2_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 3,
-                }
-            }
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # We should see events before the ban but not after
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                event_response3["event_id"],
-                event_response4["event_id"],
-                user1_ban_response["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-        # No "live" events in an initial sync (no `from_token` to define the "live"
-        # range)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            0,
-            response_body["rooms"][room_id1],
-        )
-        # There are more events to paginate to
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            True,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_rooms_ban_incremental_sync1(self) -> None:
-        """
-        Test that `rooms` we are banned from during the next incremental sync only
-        allows us to see timeline events up to the ban event.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity before1", tok=user2_tok)
-        self.helper.send(room_id1, "activity before2", tok=user2_tok)
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 4,
-                }
-            }
-        }
-        _, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        event_response3 = self.helper.send(room_id1, "activity after3", tok=user2_tok)
-        event_response4 = self.helper.send(room_id1, "activity after4", tok=user2_tok)
-        # The ban is within the token range (between the `from_token` and the sliding
-        # sync request)
-        user1_ban_response = self.helper.ban(
-            room_id1, src=user2_id, targ=user1_id, tok=user2_tok
-        )
-
-        self.helper.send(room_id1, "activity after5", tok=user2_tok)
-        self.helper.send(room_id1, "activity after6", tok=user2_tok)
-
-        # Make the incremental Sliding Sync request
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        # We should see events before the ban but not after
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                event_response3["event_id"],
-                event_response4["event_id"],
-                user1_ban_response["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-        # All live events in the incremental sync
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            3,
-            response_body["rooms"][room_id1],
-        )
-        # There aren't anymore events to paginate to in this range
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            False,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_rooms_ban_incremental_sync2(self) -> None:
-        """
-        Test that `rooms` we are banned from before the incremental sync don't return
-        any events in the timeline.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send(room_id1, "activity before1", tok=user2_tok)
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        self.helper.send(room_id1, "activity after2", tok=user2_tok)
-        # The ban is before we get our `from_token`
-        self.helper.ban(room_id1, src=user2_id, targ=user1_id, tok=user2_tok)
-
-        self.helper.send(room_id1, "activity after3", tok=user2_tok)
-
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 4,
-                }
-            }
-        }
-        _, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        self.helper.send(room_id1, "activity after4", tok=user2_tok)
-
-        # Make the incremental Sliding Sync request
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        # Nothing to see for this banned user in the room in the token range
-        self.assertIsNone(response_body["rooms"].get(room_id1))
-
-    def test_room_subscriptions_with_join_membership(self) -> None:
-        """
-        Test `room_subscriptions` with a joined room should give us timeline and current
-        state events.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        join_response = self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        # Make the Sliding Sync request with just the room subscription
-        sync_body = {
-            "room_subscriptions": {
-                room_id1: {
-                    "required_state": [
-                        [EventTypes.Create, ""],
-                    ],
-                    "timeline_limit": 1,
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        state_map = self.get_success(
-            self.storage_controllers.state.get_current_state(room_id1)
-        )
-
-        # We should see some state
-        self._assertRequiredStateIncludes(
-            response_body["rooms"][room_id1]["required_state"],
-            {
-                state_map[(EventTypes.Create, "")],
-            },
-            exact=True,
-        )
-        self.assertIsNone(response_body["rooms"][room_id1].get("invite_state"))
-
-        # We should see some events
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                join_response["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-        # No "live" events in an initial sync (no `from_token` to define the "live"
-        # range)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            0,
-            response_body["rooms"][room_id1],
-        )
-        # There are more events to paginate to
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            True,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_room_subscriptions_with_leave_membership(self) -> None:
-        """
-        Test `room_subscriptions` with a leave room should give us timeline and state
-        events up to the leave event.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.send_state(
-            room_id1,
-            event_type="org.matrix.foo_state",
-            state_key="",
-            body={"foo": "bar"},
-            tok=user2_tok,
-        )
-
-        join_response = self.helper.join(room_id1, user1_id, tok=user1_tok)
-        leave_response = self.helper.leave(room_id1, user1_id, tok=user1_tok)
-
-        state_map = self.get_success(
-            self.storage_controllers.state.get_current_state(room_id1)
-        )
-
-        # Send some events after user1 leaves
-        self.helper.send(room_id1, "activity after leave", tok=user2_tok)
-        # Update state after user1 leaves
-        self.helper.send_state(
-            room_id1,
-            event_type="org.matrix.foo_state",
-            state_key="",
-            body={"foo": "qux"},
-            tok=user2_tok,
-        )
-
-        # Make the Sliding Sync request with just the room subscription
-        sync_body = {
-            "room_subscriptions": {
-                room_id1: {
-                    "required_state": [
-                        ["org.matrix.foo_state", ""],
-                    ],
-                    "timeline_limit": 2,
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # We should see the state at the time of the leave
-        self._assertRequiredStateIncludes(
-            response_body["rooms"][room_id1]["required_state"],
-            {
-                state_map[("org.matrix.foo_state", "")],
-            },
-            exact=True,
-        )
-        self.assertIsNone(response_body["rooms"][room_id1].get("invite_state"))
-
-        # We should see some before we left (nothing after)
-        self.assertEqual(
-            [
-                event["event_id"]
-                for event in response_body["rooms"][room_id1]["timeline"]
-            ],
-            [
-                join_response["event_id"],
-                leave_response["event_id"],
-            ],
-            response_body["rooms"][room_id1]["timeline"],
-        )
-        # No "live" events in an initial sync (no `from_token` to define the "live"
-        # range)
-        self.assertEqual(
-            response_body["rooms"][room_id1]["num_live"],
-            0,
-            response_body["rooms"][room_id1],
-        )
-        # There are more events to paginate to
-        self.assertEqual(
-            response_body["rooms"][room_id1]["limited"],
-            True,
-            response_body["rooms"][room_id1],
-        )
-
-    def test_room_subscriptions_no_leak_private_room(self) -> None:
-        """
-        Test `room_subscriptions` with a private room we have never been in should not
-        leak any data to the user.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok, is_public=False)
-
-        # We should not be able to join the private room
-        self.helper.join(
-            room_id1, user1_id, tok=user1_tok, expect_code=HTTPStatus.FORBIDDEN
-        )
-
-        # Make the Sliding Sync request with just the room subscription
-        sync_body = {
-            "room_subscriptions": {
-                room_id1: {
-                    "required_state": [
-                        [EventTypes.Create, ""],
-                    ],
-                    "timeline_limit": 1,
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # We should not see the room at all (we're not in it)
-        self.assertIsNone(response_body["rooms"].get(room_id1), response_body["rooms"])
-
-    def test_room_subscriptions_world_readable(self) -> None:
-        """
-        Test `room_subscriptions` with a room that has `world_readable` history visibility
-
-        FIXME: We should be able to see the room timeline and state
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        # Create a room with `world_readable` history visibility
-        room_id1 = self.helper.create_room_as(
-            user2_id,
-            tok=user2_tok,
-            extra_content={
-                "preset": "public_chat",
-                "initial_state": [
-                    {
-                        "content": {
-                            "history_visibility": HistoryVisibility.WORLD_READABLE
-                        },
-                        "state_key": "",
-                        "type": EventTypes.RoomHistoryVisibility,
-                    }
-                ],
-            },
-        )
-        # Ensure we're testing with a room with `world_readable` history visibility
-        # which means events are visible to anyone even without membership.
-        history_visibility_response = self.helper.get_state(
-            room_id1, EventTypes.RoomHistoryVisibility, tok=user2_tok
-        )
-        self.assertEqual(
-            history_visibility_response.get("history_visibility"),
-            HistoryVisibility.WORLD_READABLE,
-        )
-
-        # Note: We never join the room
-
-        # Make the Sliding Sync request with just the room subscription
-        sync_body = {
-            "room_subscriptions": {
-                room_id1: {
-                    "required_state": [
-                        [EventTypes.Create, ""],
-                    ],
-                    "timeline_limit": 1,
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # FIXME: In the future, we should be able to see the room because it's
-        # `world_readable` but currently we don't support this.
-        self.assertIsNone(response_body["rooms"].get(room_id1), response_body["rooms"])
-
-    # Any extensions that use `lists`/`rooms` should be tested here
-    @parameterized.expand([("account_data",), ("receipts",)])
-    def test_extensions_lists_rooms_relevant_rooms(self, extension_name: str) -> None:
-        """
-        With various extensions, test out requesting different variations of
-        `lists`/`rooms`.
-        """
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        # Create some rooms
-        room_id1 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id2 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id3 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id4 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id5 = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        room_id_to_human_name_map = {
-            room_id1: "room1",
-            room_id2: "room2",
-            room_id3: "room3",
-            room_id4: "room4",
-            room_id5: "room5",
-        }
-
-        for room_id in room_id_to_human_name_map.keys():
-            if extension_name == "account_data":
-                # Add some account data to each room
-                self.get_success(
-                    self.account_data_handler.add_account_data_to_room(
-                        user_id=user1_id,
-                        room_id=room_id,
-                        account_data_type="org.matrix.roorarraz",
-                        content={"roo": "rar"},
-                    )
-                )
-            elif extension_name == "receipts":
-                event_response = self.helper.send(
-                    room_id, body="new event", tok=user1_tok
-                )
-                # Read last event
-                channel = self.make_request(
-                    "POST",
-                    f"/rooms/{room_id}/receipt/{ReceiptTypes.READ}/{event_response['event_id']}",
-                    {},
-                    access_token=user1_tok,
-                )
-                self.assertEqual(channel.code, 200, channel.json_body)
-            else:
-                raise AssertionError(f"Unknown extension name: {extension_name}")
-
-        main_sync_body = {
-            "lists": {
-                # We expect this list range to include room5 and room4
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                },
-                # We expect this list range to include room5, room4, room3
-                "bar-list": {
-                    "ranges": [[0, 2]],
-                    "required_state": [],
-                    "timeline_limit": 0,
-                },
-            },
-            "room_subscriptions": {
-                room_id1: {
-                    "required_state": [],
-                    "timeline_limit": 0,
-                }
-            },
-        }
-
-        # Mix lists and rooms
-        sync_body = {
-            **main_sync_body,
-            "extensions": {
-                extension_name: {
-                    "enabled": True,
-                    "lists": ["foo-list", "non-existent-list"],
-                    "rooms": [room_id1, room_id2, "!non-existent-room"],
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # room1: ✅ Requested via `rooms` and a room subscription exists
-        # room2: ❌ Requested via `rooms` but not in the response (from lists or room subscriptions)
-        # room3: ❌ Not requested
-        # room4: ✅ Shows up because requested via `lists` and list exists in the response
-        # room5: ✅ Shows up because requested via `lists` and list exists in the response
-        self.assertIncludes(
-            {
-                room_id_to_human_name_map[room_id]
-                for room_id in response_body["extensions"][extension_name]
-                .get("rooms")
-                .keys()
-            },
-            {"room1", "room4", "room5"},
-            exact=True,
-        )
-
-        # Try wildcards (this is the default)
-        sync_body = {
-            **main_sync_body,
-            "extensions": {
-                extension_name: {
-                    "enabled": True,
-                    # "lists": ["*"],
-                    # "rooms": ["*"],
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # room1: ✅ Shows up because of default `rooms` wildcard and is in one of the room subscriptions
-        # room2: ❌ Not requested
-        # room3: ✅ Shows up because of default `lists` wildcard and is in a list
-        # room4: ✅ Shows up because of default `lists` wildcard and is in a list
-        # room5: ✅ Shows up because of default `lists` wildcard and is in a list
-        self.assertIncludes(
-            {
-                room_id_to_human_name_map[room_id]
-                for room_id in response_body["extensions"][extension_name]
-                .get("rooms")
-                .keys()
-            },
-            {"room1", "room3", "room4", "room5"},
-            exact=True,
-        )
-
-        # Empty list will return nothing
-        sync_body = {
-            **main_sync_body,
-            "extensions": {
-                extension_name: {
-                    "enabled": True,
-                    "lists": [],
-                    "rooms": [],
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # room1: ❌ Not requested
-        # room2: ❌ Not requested
-        # room3: ❌ Not requested
-        # room4: ❌ Not requested
-        # room5: ❌ Not requested
-        self.assertIncludes(
-            {
-                room_id_to_human_name_map[room_id]
-                for room_id in response_body["extensions"][extension_name]
-                .get("rooms")
-                .keys()
-            },
-            set(),
-            exact=True,
-        )
-
-        # Try wildcard and none
-        sync_body = {
-            **main_sync_body,
-            "extensions": {
-                extension_name: {
-                    "enabled": True,
-                    "lists": ["*"],
-                    "rooms": [],
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # room1: ❌ Not requested
-        # room2: ❌ Not requested
-        # room3: ✅ Shows up because of default `lists` wildcard and is in a list
-        # room4: ✅ Shows up because of default `lists` wildcard and is in a list
-        # room5: ✅ Shows up because of default `lists` wildcard and is in a list
-        self.assertIncludes(
-            {
-                room_id_to_human_name_map[room_id]
-                for room_id in response_body["extensions"][extension_name]
-                .get("rooms")
-                .keys()
-            },
-            {"room3", "room4", "room5"},
-            exact=True,
-        )
-
-        # Try requesting a room that is only in a list
-        sync_body = {
-            **main_sync_body,
-            "extensions": {
-                extension_name: {
-                    "enabled": True,
-                    "lists": [],
-                    "rooms": [room_id5],
-                }
-            },
-        }
-        response_body, _ = self.do_sync(sync_body, tok=user1_tok)
-
-        # room1: ❌ Not requested
-        # room2: ❌ Not requested
-        # room3: ❌ Not requested
-        # room4: ❌ Not requested
-        # room5: ✅ Requested via `rooms` and is in a list
-        self.assertIncludes(
-            {
-                room_id_to_human_name_map[room_id]
-                for room_id in response_body["extensions"][extension_name]
-                .get("rooms")
-                .keys()
-            },
-            {"room5"},
-            exact=True,
-        )
-
-    def test_rooms_required_state_incremental_sync_LIVE(self) -> None:
-        """Test that we only get state updates in incremental sync for rooms
-        we've already seen (LIVE).
-        """
-
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-        user2_id = self.register_user("user2", "pass")
-        user2_tok = self.login(user2_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user2_id, tok=user2_tok)
-        self.helper.join(room_id1, user1_id, tok=user1_tok)
-
-        # Make the Sliding Sync request
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 1]],
-                    "required_state": [
-                        [EventTypes.Create, ""],
-                        [EventTypes.RoomHistoryVisibility, ""],
-                        # This one doesn't exist in the room
-                        [EventTypes.Name, ""],
-                    ],
-                    "timeline_limit": 0,
-                }
-            }
-        }
-
-        response_body, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        state_map = self.get_success(
-            self.storage_controllers.state.get_current_state(room_id1)
-        )
-
-        self._assertRequiredStateIncludes(
-            response_body["rooms"][room_id1]["required_state"],
-            {
-                state_map[(EventTypes.Create, "")],
-                state_map[(EventTypes.RoomHistoryVisibility, "")],
-            },
-            exact=True,
-        )
-
-        # Send a state event
-        self.helper.send_state(
-            room_id1, EventTypes.Name, body={"name": "foo"}, tok=user2_tok
-        )
-
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        state_map = self.get_success(
-            self.storage_controllers.state.get_current_state(room_id1)
-        )
-
-        self.assertNotIn("initial", response_body["rooms"][room_id1])
-        self._assertRequiredStateIncludes(
-            response_body["rooms"][room_id1]["required_state"],
-            {
-                state_map[(EventTypes.Name, "")],
-            },
-            exact=True,
-        )
-
-    @parameterized.expand([(False,), (True,)])
-    def test_rooms_timeline_incremental_sync_PREVIOUSLY(self, limited: bool) -> None:
-        """
-        Test getting room data where we have previously sent down the room, but
-        we missed sending down some timeline events previously and so its status
-        is considered PREVIOUSLY.
-
-        There are two versions of this test, one where there are more messages
-        than the timeline limit, and one where there isn't.
-        """
-
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id2 = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        self.helper.send(room_id1, "msg", tok=user1_tok)
-
-        timeline_limit = 5
-        conn_id = "conn_id"
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 0]],
-                    "required_state": [],
-                    "timeline_limit": timeline_limit,
-                }
-            },
-            "conn_id": "conn_id",
-        }
-
-        # The first room gets sent down the initial sync
-        response_body, initial_from_token = self.do_sync(sync_body, tok=user1_tok)
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id1}, response_body["rooms"]
-        )
-
-        # We now send down some events in room1 (depending on the test param).
-        expected_events = []  # The set of events in the timeline
-        if limited:
-            for _ in range(10):
-                resp = self.helper.send(room_id1, "msg1", tok=user1_tok)
-                expected_events.append(resp["event_id"])
-        else:
-            resp = self.helper.send(room_id1, "msg1", tok=user1_tok)
-            expected_events.append(resp["event_id"])
-
-        # A second messages happens in the other room, so room1 won't get sent down.
-        self.helper.send(room_id2, "msg", tok=user1_tok)
-
-        # Only the second room gets sent down sync.
-        response_body, from_token = self.do_sync(
-            sync_body, since=initial_from_token, tok=user1_tok
-        )
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id2}, response_body["rooms"]
-        )
-
-        # FIXME: This is a hack to record that the first room wasn't sent down
-        # sync, as we don't implement that currently.
-        sliding_sync_handler = self.hs.get_sliding_sync_handler()
-        requester = self.get_success(
-            self.hs.get_auth().get_user_by_access_token(user1_tok)
-        )
-        sync_config = SlidingSyncConfig(
-            user=requester.user,
-            requester=requester,
-            conn_id=conn_id,
-        )
-
-        parsed_initial_from_token = self.get_success(
-            SlidingSyncStreamToken.from_string(self.store, initial_from_token)
-        )
-        connection_position = self.get_success(
-            sliding_sync_handler.connection_store.record_rooms(
-                sync_config,
-                parsed_initial_from_token,
-                sent_room_ids=[],
-                unsent_room_ids=[room_id1],
-            )
-        )
-
-        # FIXME: Now fix up `from_token` with new connect position above.
-        parsed_from_token = self.get_success(
-            SlidingSyncStreamToken.from_string(self.store, from_token)
-        )
-        parsed_from_token = SlidingSyncStreamToken(
-            stream_token=parsed_from_token.stream_token,
-            connection_position=connection_position,
-        )
-        from_token = self.get_success(parsed_from_token.to_string(self.store))
-
-        # We now send another event to room1, so we should sync all the missing events.
-        resp = self.helper.send(room_id1, "msg2", tok=user1_tok)
-        expected_events.append(resp["event_id"])
-
-        # This sync should contain the messages from room1 not yet sent down.
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id1}, response_body["rooms"]
-        )
-        self.assertNotIn("initial", response_body["rooms"][room_id1])
-
-        self.assertEqual(
-            [ev["event_id"] for ev in response_body["rooms"][room_id1]["timeline"]],
-            expected_events[-timeline_limit:],
-        )
-        self.assertEqual(response_body["rooms"][room_id1]["limited"], limited)
-        self.assertEqual(response_body["rooms"][room_id1].get("required_state"), None)
-
-    def test_rooms_required_state_incremental_sync_PREVIOUSLY(self) -> None:
-        """
-        Test getting room data where we have previously sent down the room, but
-        we missed sending down some state previously and so its status is
-        considered PREVIOUSLY.
-        """
-
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id2 = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        self.helper.send(room_id1, "msg", tok=user1_tok)
-
-        conn_id = "conn_id"
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 0]],
-                    "required_state": [
-                        [EventTypes.Create, ""],
-                        [EventTypes.RoomHistoryVisibility, ""],
-                        # This one doesn't exist in the room
-                        [EventTypes.Name, ""],
-                    ],
-                    "timeline_limit": 0,
-                }
-            },
-            "conn_id": "conn_id",
-        }
-
-        # The first room gets sent down the initial sync
-        response_body, initial_from_token = self.do_sync(sync_body, tok=user1_tok)
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id1}, response_body["rooms"]
-        )
-
-        # We now send down some state in room1
-        resp = self.helper.send_state(
-            room_id1, EventTypes.Name, {"name": "foo"}, tok=user1_tok
-        )
-        name_change_id = resp["event_id"]
-
-        # A second messages happens in the other room, so room1 won't get sent down.
-        self.helper.send(room_id2, "msg", tok=user1_tok)
-
-        # Only the second room gets sent down sync.
-        response_body, from_token = self.do_sync(
-            sync_body, since=initial_from_token, tok=user1_tok
-        )
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id2}, response_body["rooms"]
-        )
-
-        # FIXME: This is a hack to record that the first room wasn't sent down
-        # sync, as we don't implement that currently.
-        sliding_sync_handler = self.hs.get_sliding_sync_handler()
-        requester = self.get_success(
-            self.hs.get_auth().get_user_by_access_token(user1_tok)
-        )
-        sync_config = SlidingSyncConfig(
-            user=requester.user,
-            requester=requester,
-            conn_id=conn_id,
-        )
-
-        parsed_initial_from_token = self.get_success(
-            SlidingSyncStreamToken.from_string(self.store, initial_from_token)
-        )
-        connection_position = self.get_success(
-            sliding_sync_handler.connection_store.record_rooms(
-                sync_config,
-                parsed_initial_from_token,
-                sent_room_ids=[],
-                unsent_room_ids=[room_id1],
-            )
-        )
-
-        # FIXME: Now fix up `from_token` with new connect position above.
-        parsed_from_token = self.get_success(
-            SlidingSyncStreamToken.from_string(self.store, from_token)
-        )
-        parsed_from_token = SlidingSyncStreamToken(
-            stream_token=parsed_from_token.stream_token,
-            connection_position=connection_position,
-        )
-        from_token = self.get_success(parsed_from_token.to_string(self.store))
-
-        # We now send another event to room1, so we should sync all the missing state.
-        self.helper.send(room_id1, "msg", tok=user1_tok)
-
-        # This sync should contain the state changes from room1.
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id1}, response_body["rooms"]
-        )
-        self.assertNotIn("initial", response_body["rooms"][room_id1])
-
-        # We should only see the name change.
-        self.assertEqual(
-            [
-                ev["event_id"]
-                for ev in response_body["rooms"][room_id1]["required_state"]
-            ],
-            [name_change_id],
-        )
-
-    def test_rooms_required_state_incremental_sync_NEVER(self) -> None:
-        """
-        Test getting `required_state` where we have NEVER sent down the room before
-        """
-
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id2 = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        self.helper.send(room_id1, "msg", tok=user1_tok)
-
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 0]],
-                    "required_state": [
-                        [EventTypes.Create, ""],
-                        [EventTypes.RoomHistoryVisibility, ""],
-                        # This one doesn't exist in the room
-                        [EventTypes.Name, ""],
-                    ],
-                    "timeline_limit": 1,
-                }
-            },
-        }
-
-        # A message happens in the other room, so room1 won't get sent down.
-        self.helper.send(room_id2, "msg", tok=user1_tok)
-
-        # Only the second room gets sent down sync.
-        response_body, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id2}, response_body["rooms"]
-        )
-
-        # We now send another event to room1, so we should send down the full
-        # room.
-        self.helper.send(room_id1, "msg2", tok=user1_tok)
-
-        # This sync should contain the messages from room1 not yet sent down.
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id1}, response_body["rooms"]
-        )
-
-        self.assertEqual(response_body["rooms"][room_id1]["initial"], True)
-
-        state_map = self.get_success(
-            self.storage_controllers.state.get_current_state(room_id1)
-        )
-
-        self._assertRequiredStateIncludes(
-            response_body["rooms"][room_id1]["required_state"],
-            {
-                state_map[(EventTypes.Create, "")],
-                state_map[(EventTypes.RoomHistoryVisibility, "")],
-            },
-            exact=True,
-        )
-
-    def test_rooms_timeline_incremental_sync_NEVER(self) -> None:
-        """
-        Test getting timeline room data where we have NEVER sent down the room
-        before
-        """
-
-        user1_id = self.register_user("user1", "pass")
-        user1_tok = self.login(user1_id, "pass")
-
-        room_id1 = self.helper.create_room_as(user1_id, tok=user1_tok)
-        room_id2 = self.helper.create_room_as(user1_id, tok=user1_tok)
-
-        sync_body = {
-            "lists": {
-                "foo-list": {
-                    "ranges": [[0, 0]],
-                    "required_state": [],
-                    "timeline_limit": 5,
-                }
-            },
-        }
-
-        expected_events = []
-        for _ in range(4):
-            resp = self.helper.send(room_id1, "msg", tok=user1_tok)
-            expected_events.append(resp["event_id"])
-
-        # A message happens in the other room, so room1 won't get sent down.
-        self.helper.send(room_id2, "msg", tok=user1_tok)
-
-        # Only the second room gets sent down sync.
-        response_body, from_token = self.do_sync(sync_body, tok=user1_tok)
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id2}, response_body["rooms"]
-        )
-
-        # We now send another event to room1 so it comes down sync
-        resp = self.helper.send(room_id1, "msg2", tok=user1_tok)
-        expected_events.append(resp["event_id"])
-
-        # This sync should contain the messages from room1 not yet sent down.
-        response_body, _ = self.do_sync(sync_body, since=from_token, tok=user1_tok)
-
-        self.assertCountEqual(
-            response_body["rooms"].keys(), {room_id1}, response_body["rooms"]
-        )
-
-        self.assertEqual(
-            [ev["event_id"] for ev in response_body["rooms"][room_id1]["timeline"]],
-            expected_events,
-        )
-        self.assertEqual(response_body["rooms"][room_id1]["limited"], True)
-        self.assertEqual(response_body["rooms"][room_id1]["initial"], True)
-
     def test_rooms_with_no_updates_do_not_come_down_incremental_sync(self) -> None:
         """
         Test that rooms with no updates are returned in subsequent incremental
